Add the Storage page, item details and a locations editor
Replace the single-list demo with the Storage page, backed by a new location
upsert endpoint, and make item quantities whole numbers end to end.

Frontend (React Router 8, Tailwind CSS 4, MobX 7):
- App shell: sidebar on wide screens, bottom tab bar on phones; Shopping and
  Planner placeholders and a Profile page
- /storage/:locationId: location tabs in stored order, search with per-location
  match counts, sort by name, expiry, quantity or size kept in the URL, cards
  three a row on phones and at most seven, selection with delete and move
- Quantity steps are the one optimistic update: an overlay in QuantityUpdates,
  sent 600 ms after the last tap, then a background refetch that keeps
  unchanged items' identity so the grid does not repaint
- /storage/:locationId/items/:itemId: full-screen details with the stepper,
  an expiry breakdown, mark as opened and remove; an edit form that sends only
  changed fields, with unsaved changes guarded by useBlocker
- Locations editor: rename, add, delete (moving items elsewhere) and reorder
  by drag and drop with @dnd-kit, saved in one PUT and rebased onto changes
  other members make meanwhile
- Native <dialog> and Popover API primitives; all copy in i18n/messages.ts,
  ready for localization
- Shopping lists and bulk item endpoints do not exist yet: add-to-list actions
  are inactive, and bulk delete and move send one request per item

Backend:
- PUT /households/:householdId/locations saves the editor in one transaction:
  renames (swaps included), additions, deletions with moveItemsTo, and order;
  409 when the edited list is stale; one LocationsUpserted broadcast
- Quantity deltas are exact integers, so the float rounding is gone

Database:
- items.quantity and item_events.quantity_delta are integers. Migration 0000
  and both snapshots were edited in place instead of adding a migration, so a
  database created before this commit must be dropped and re-created
- Test fixture: Flour is 1 pcs of 1.5 kg rather than 1.5 kg

Shared:
- Quantity is validated with @IsInt(); QUANTITY_DECIMAL_PLACES becomes
  SIZE_DECIMAL_PLACES, since only sizes keep three decimals
- Location upsert DTOs, the LocationsUpserted event and its payload

+++ packages/db/migrations/0000_round_absorbing_man.sql
@@ -85,7 +85,7 @@ CREATE TABLE "items" (
 	"name" varchar(80) NOT NULL,
 	"location_id" uuid NOT NULL,
 	"category" text NOT NULL,
-	"quantity" numeric(10, 3) NOT NULL,
+	"quantity" integer NOT NULL,
 	"unit" text NOT NULL,
 	"size_value" numeric(10, 3),
 	"size_unit" text,
@@ -111,7 +111,7 @@ CREATE TABLE "item_events" (
 	"item_id" uuid NOT NULL,
 	"user_id" uuid,
 	"type" text NOT NULL,
-	"quantity_delta" numeric(10, 3),
+	"quantity_delta" integer,
 	"payload" jsonb DEFAULT '{}'::jsonb NOT NULL,
 	"created_at" timestamp with time zone DEFAULT now() NOT NULL,
 	CONSTRAINT "item_events_type_check" CHECK (type in ('added', 'updated', 'opened', 'consumed', 'discarded', 'restored'))
